a bit more optimization + organize files
