+++ src/commands/stats.ts
@@ -1,7 +1,6 @@
 import type { UnknownString } from "@zakahacecosas/string-utils";
 import { LogStuff } from "../functions/io.ts";
 import { GetProjectEnvironment } from "../functions/projects.ts";
-import { NameProject } from "../functions/projects.ts";
 import type { FnCPF } from "../types/platform.ts";
 import { RecommendedCommunityStandards } from "./toolkit/rcs.ts";
 import { ColorString } from "../functions/color.ts";
@@ -22,10 +21,9 @@ function StringifyDependencyRelationship(rel: FnCPF["deps"][0]["rel"]): string {
 
 export default async function TheStatistics(target: UnknownString): Promise<void> {
     const env = await GetProjectEnvironment(target);
-    const name = await NameProject(env.root, "all");
 
     LogStuff(
-        `${name} · ${ColorString(env.runtime, "bold")} runtime · ${ColorString(env.manager, "bold")} pkg manager`,
+        `${env.names.full}\n${ColorString(env.runtime, "bold")} runtime · ${ColorString(env.manager, "bold")} pkg manager`,
     );
 
     const maxDeps = 3;
