+++ src/commands/updater.ts
@@ -1,14 +1,14 @@
 import { compare, parse } from "@std/semver";
 import type { GITHUB_RELEASE } from "../types/misc.ts";
 import { GetDateNow } from "../functions/date.ts";
-import type { TheUpdaterConstructedParams } from "./constructors/command.ts";
+import type { TheUpdaterConstructedParams } from "./_interfaces.ts";
 import { LogStuff, StringifyYaml } from "../functions/io.ts";
 import { parse as parseYaml } from "@std/yaml";
 import { GetAppPath } from "../functions/config.ts";
 import type { CF_FKNODE_SCHEDULE } from "../types/config_files.ts";
 import * as DenoJson from "../../deno.json" with { type: "json" };
 import { ColorString } from "../functions/color.ts";
-import { LOCAL_PLATFORM } from "../constants/platform.ts";
+import { LOCAL_PLATFORM } from "../platform.ts";
 import { validate } from "@zakahacecosas/string-utils";
 
 async function CheckUpdates(): Promise<CF_FKNODE_SCHEDULE | "rl"> {
