+++ src/platform.ts
