+++ tests/environment/test-one/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "uwu.js",
+    "name": "a.js",
     "version": "1.0.0",
     "description": "Certainly! Here's a description for a useless NPM package called 'uwu.js': \"A simple package to uwu-ify your text.\"",
     "main": "index.js",
